complete with write

+++ writeDisk.h
@@ -13,16 +13,16 @@ namespace cachemanager
 {
 class DiskManager{
 public:
-	bool writeOneBuff(const string & hashName,const char buff[],const unsigned int &offset,int &length);
+	bool writeOneBuff(const string & hashName,const char buff[],const unsigned int &offset,unsigned int &length);
   static DiskManager *getInstance(const size_t &ds);
 private:
   /*begin 和end 通过计算得到的跨越块的情况*/
   //unsigned int begin;
   //unsigned int end;
-  const unsigned long int   diskSize;
+   const  unsigned int   diskSize;
 	DiskManager(const size_t &ds):diskSize(ds*1024*1024)
   {
-
+     //cout<<"ds="<<ds<<endl;
 	}
   void freeNew(char * str);
   //static DiskManager * m_Instance;
